Address review comments

+++ commercial/Boltdir/modules/stack/tasks/create_code.sh
@@ -7,7 +7,10 @@ die() {
 
 cd pupperware || die no-repo "run the clone task first to set up pupperware"
 
-rm -rf code/*
+if [[ -d code/ ]]; then
+    sudo chown -R "$USER" code/
+    rm -rf code/*
+fi
 mkdir -p code/environments/production/manifests
 cat > code/environments/production/manifests/site.pp <<'EOF'
 $timestamp = strftime('%Y-%m-%dT%H:%M:%S')
